+++ src/jsc/jsc_types.hpp
@@ -30,7 +30,6 @@ namespace jsc {
 struct Types {
     using Context = JSContextRef;
     using GlobalContext = JSGlobalContextRef;
-    using ClassDefinition = JSClassRef;
     using Value = JSValueRef;
     using Object = JSObjectRef;
     using String = JSStringRef;