+++ CHANGELOG.md
@@ -5,9 +5,10 @@ x.x.x Release notes (yyyy-MM-dd)
 
 ### Enhancements
 * Added `isValid()` method to `List` and `Results` to check for deleleted or invalidated objects
+* Added `objectForPrimaryKey(type, key)` method to `Realm`
 
 ### Bugfixes
-* None
+* Fix for crash when setting object properties to objects from other Realms
 
 0.13.2 Release notes (2016-5-26)
 =============================================================